+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -308,6 +308,7 @@ object HybrisConstants {
     const val CLASS_NAME_ENUM = "HybrisEnumValue"
     const val CLASS_FQN_INTERCEPTOR_TYPE = "de.hybris.platform.servicelayer.interceptor.impl.InterceptorExecutionPolicy.InterceptorType"
     const val CLASS_FQN_INTERCEPTOR_MAPPING = "de.hybris.platform.servicelayer.interceptor.impl.InterceptorMapping"
+    const val CLASS_FQN_CMS_RESTRICTION_EVALUATOR_MAPPING = "de.hybris.platform.cms2.servicelayer.services.evaluator.CMSRestrictionEvaluatorMapping"
     const val CLASS_FQN_ANNOTATION_ACCESSOR = "de.hybris.bootstrap.annotations.Accessor"
     const val CLASS_FQN_CONFIG_IMPORT_PROCESSOR = "de.hybris.platform.commerceservices.impex.impl.ConfigPropertyImportProcessor"
     const val CLASS_FQN_CONVERTER = "de.hybris.platform.servicelayer.dto.converter.Converter"