+++ src/com/intellij/idea/plugin/hybris/system/type/codeInsight/completion/TSCompletionContributor.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -35,7 +35,7 @@ class TSCompletionContributor : CompletionContributor() {
         )
         extend(
             CompletionType.BASIC,
-            PlatformPatterns.psiElement().inside(TSPatterns.SPRING_INTERCEPTOR_TYPE_CODE),
+            PlatformPatterns.psiElement().inside(TSPatterns.SPRING_TYPE_CODE),
             ItemTypeCodeCompletionProvider()
         )
     }