+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2024.3.5]
+
+### `Type System` enhancements
+- Added completion and navigation for `restrictionTypeCode` property of the `CMSRestrictionEvaluatorMapping` within the Spring XML file [#1300](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1300)
+
 ## [2024.3.4]
 
 ### `Integrations` enhancements